// templates/api/models/template.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DocumentType } from '../models/document-type';
import { TemplateCategory } from '../models/template-category';
import { TemplateStatus } from '../models/template-status';
export interface Template {
  category?: TemplateCategory;
  code: string;
  createdBy?: string;
  createdDate?: string;
  description?: string;
  documentType?: DocumentType;
  endDate?: string;
  modifiedBy?: string;
  modifiedDate?: string;
  name?: string;
  path?: string;
  startDate?: string;
  status: TemplateStatus;
  templateId: string;
  tenantId: string;
}